# [libc++] Optimize __hash_table::erase(iterator, iterator)